Start deleting swing testing utils covered by the library
