+++ build.gradle
@@ -7,6 +7,7 @@ repositories {
     jcenter()
 
     maven { url 'https://jitpack.io' }
+    maven { url 'https://dl.bintray.com/alexburlton/swing-test' }
 }
 
 dependencies {
@@ -24,6 +25,7 @@ dependencies {
     testImplementation "io.mockk:mockk:1.9.3"
     testImplementation "io.kotlintest:kotlintest-assertions:3.4.2"
     testImplementation 'org.jetbrains.kotlin:kotlin-test-junit'
+    testImplementation 'com.github.alexburlton:swing-test:0.1.0'
 }
 
 compileKotlin {
