+++ src/test/kotlin/dartzee/db/TestPlayerEntity.kt
@@ -6,7 +6,7 @@ import dartzee.core.util.DateStatics
 import dartzee.core.util.getSqlDateNow
 import dartzee.helper.insertPlayer
 import dartzee.helper.insertPlayerImage
-import dartzee.helper.shouldMatch
+import dartzee.shouldMatch
 import io.kotlintest.matchers.collections.shouldContainExactly
 import io.kotlintest.matchers.string.shouldNotBeEmpty
 import io.kotlintest.matchers.types.shouldBeInstanceOf
