+++ src/test/kotlin/dartzee/screen/game/scorer/TestDartsScorerDartzee.kt
@@ -5,11 +5,11 @@ import dartzee.core.helper.verifyNotCalled
 import dartzee.core.util.DateStatics
 import dartzee.core.util.getSqlDateNow
 import dartzee.dartzee.DartzeeRoundResult
-import dartzee.doClick
 import dartzee.helper.AbstractTest
 import dartzee.helper.insertGame
 import dartzee.helper.insertPlayer
 import dartzee.screen.game.dartzee.GamePanelDartzee
+import doClick
 import io.kotlintest.matchers.types.shouldBeInstanceOf
 import io.kotlintest.shouldBe
 import io.mockk.every
